requirements.txt
python-dotenv
openai
langchain
PyPDF2
streamlit
-e .